backend/requirements.txt
fastapi
uvicorn
motor
pymongo
python-multipart
python-dotenv
ultralytics
torch
torchvision
Pillow
